Convert US Healthcare Map to React application
The existing was was refactored into a React app, using the same
semantic structure, classes, and cartographic logic. Some updates and
refactoring were required to gain compatiability with the React
rendering system.

Current known outstanding issues are:
* Recent commits that adjust stying and copy need to be ported
* There is no map hover interactivity for attribute popup
* There is no legend

Tiles and break configuration are required to end up in the `public`
directory through a CI build process or similar.