Pin Dockerfile version to Python 3.8.6
There was a recent release of Python 3.9, which bumped the `python:3`
base image tag. This has caused issues with installations since
libraries like scipy and numpy haven't published wheels for this
python version yet. This commit changes the Dockerfile to be based on
the more specific 3.8.6 tag, which won't update in the future to a new
version of Python and potentially break things.